# backend/internal/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoMD API
  description: API for managing a self-hosted markdown knowledge base with real-time SSE updates and MCP integration.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Default local development server
paths:
  /api/files:
    get:
      summary: List all files
      responses:
        '200':
          description: Successful response
    post:
      summary: Create a new file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  description: Path to the new file relative to vault root
                content:
                  type: string
                  description: File content
      responses:
        '201':
          description: File created
...
  /api/folders:
    get:
      summary: List root directory contents
      responses:
        '200':
          description: Returns files and directories
    post:
      summary: Create a new directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
      responses:
        '201':
          description: Directory created
...
  /api/rename:
    post:
      summary: Rename or move a file/directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPath:
                  type: string
                newPath:
                  type: string
      responses:
        '200':
          description: Renamed successfully
  /api/search:
    get:
      summary: Search the vault
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
...
  /api/locks:
    get:
      summary: Get all file/folder locks
      responses:
        '200':
          description: Map of locked paths
    post:
      summary: Lock or unlock a path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                locked:
                  type: boolean
      responses:
        '200':
          description: Lock status updated
...